refactor(incidentio): move configuration type into incidentio package
